training/requirements.txt
tensorflow_hub==0.4.0
tensorflow_datasets
numpy
pillow